requirements.txt
groq==0.9.0
python-dotenv==1.0.1
streamlit==1.35.0